refactor: rename hw_integ_test.py
